+++ src/lm/diff.cpp
@@ -23,12 +23,12 @@ diff::diff(const cpptoml::table& config) : lm_{config}
     auto nval = table->get_as<int64_t>("n-value");
     if (!nval)
         throw diff_exception{"n-value not specified in config"};
-    n_val_ = *nval;
+    n_val_ = static_cast<uint64_t>(*nval);
 
     auto edits = table->get_as<int64_t>("max-edits");
     if (!edits)
         throw diff_exception{"max-edits not specified in config"};
-    max_edits_ = *edits;
+    max_edits_ = static_cast<uint64_t>(*edits);
 
     auto lambda = table->get_as<double>("lambda");
     lambda_ = lambda ? *lambda : 0.5;
@@ -40,7 +40,8 @@ diff::diff(const cpptoml::table& config) : lm_{config}
     substitute_penalty_
         = table->get_as<double>("substitute-penalty").value_or(0.0);
     remove_penalty_ = table->get_as<double>("remove-penalty").value_or(0.0);
-    max_cand_size_ = table->get_as<int64_t>("max-candidates").value_or(20);
+    max_cand_size_ = static_cast<uint16_t>(
+        table->get_as<int64_t>("max-candidates").value_or(20));
     lm_generate_ = table->get_as<bool>("lm-generate").value_or(false);
 
     set_stems(*table);
@@ -48,7 +49,7 @@ diff::diff(const cpptoml::table& config) : lm_{config}
 }
 
 std::vector<std::pair<sentence, double>>
-    diff::candidates(const sentence& sent, bool use_lm /* = false */)
+diff::candidates(const sentence& sent, bool use_lm /* = false */)
 {
     use_lm_ = use_lm;
     using pair_t = std::pair<sentence, double>;
